Fix five code review issues: docs, dead code, truncation, init order, round-trip
- Fix AES padding docs in crypto_primitives.hpp: said "no PKCS7 padding"
  but both FC's aes_encrypt() and SubtleCrypto use PKCS#7. Updated to
  match the actual behavior.

- Remove dead salt parameter from core_minimal's create_session() and
  generate_token() — tokens are generated from crypto random bytes and
  the salt was never read. WASM API still accepts salt for compat.

- Fix nonce truncation at WASM boundary: changed encrypt_data's nonce
  from uint32_t to double. JS number (IEEE 754) preserves integers up
  to 2^53, sufficient for microsecond timestamps that uint32_t silently
  truncated.

- Fix fragile C++ initialization order in wasm_crypto_provider using
  base-from-member idiom: prims_ now lives in a private base class that
  initializes before crypto_provider_impl binds a reference to it.

- Eliminate JSON round-trip in encrypt_and_save(): split out
  encrypt_wallet_keys() returning raw cipher bytes, avoiding the
  encrypt→hex→JSON→parse→hex-decode cycle. Added it to ASYNCIFY_ONLY.

- Remove five dead Asyncify patterns (beekeeper::*) that no longer
  match any symbols in the WASM binary.

+++ programs/beekeeper/beekeeper_wasm/include/beekeeper_wasm/beekeeper_wasm_api.hpp
@@ -72,7 +72,7 @@ public:
 
   std::string encrypt_data(const std::string& token, const std::string& wallet_name,
                            const std::string& from_key, const std::string& to_key,
-                           const std::string& content, uint32_t nonce);
+                           const std::string& content, double nonce);
   std::string decrypt_data(const std::string& token, const std::string& wallet_name,
                            const std::string& from_key, const std::string& to_key,
                            const std::string& encrypted_content);
